+++ BigoAds/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 5.1.0.0
+* Certified with BigoAds SDK 5.1.0.
+
 ## 5.0.6.0
 * Certified with BigoAds SDK 5.0.6.