+++ Line/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 3.0.1.0
+* Certified with Line SDK 3.0.1.
+
+## 3.0.0.0
+* Certified with Line SDK 3.0.0.
+
 ## 2.9.20251119.0
 * Certified with Line SDK 2.9.20251119.